Fix building with ninja build
Unsure of if it's a bug in cmake or ninja, but when using the cmake
Ninja generator, the dependency ordering places the _link_ step of the
test executable after the link step of externis.so. This means that
ninja attempts to compile test.cc into test.o gets ordered before
externis.so exists, which obviously doesn't work.

The cmake documentation specifies the DEPENDS in a custom target can
only depend on files created in the directory where the custom target is
declared, hence adding this new dependency rule in the root CMakeLists
file.