Support targeting GCC other than the one being used to build externis (#8)
* Use GCC_PLUGIN_INCLUDE if pre-defined

This makes it possible to compile externis for a cross-compilation
toolchain, where the plugin directory for the GCC used to build externis
is not the plugin folder of the GCC that externis will be used with.

* Add cmake option to disable the build test

When cross-compiling externis for another toolchain, it may not be
possible to build the test case.

* Rename GCC_PLUGIN_INCLUDE to EXTERNIS_GCC_PLUGIN_DIR

"_DIR" better represents what the purpose of the directory is, and
prefixing with EXTERNIS makes it clear that this is a project-specific
setting and not some special CMake or environmental thing.

* Add README section about cross-compilation

* Fix dependency ordering issue with GNU Make